Set default values for AlertNrqlCondition on deserialization if absent
The API response for the creation of an AlertNrqlCondition with a type
other than the default (static), such as an outlier condition, does not
contain the `value_function` JSON field, and thus the struct's
`ValueFunction` is set to the default string value (i.e. an empty
string).

That discrepancy creates a (virtual) difference between the plan and the
state which can never be resolved, and thus `terraform apply` would seem
to fail despite the changes being correctly applied.

The approach taken here is setting default values for absent fields when
they're absent from the response, which maintains backward compatibility
with existing plans.

An alternative approach where the default value was removed from the
field definition in AlertNrqlCondition was also considered but discarded
because it created a breaking change.