fix(alert_condition): mark condition_scope optional
resolves: #386

Without this change, the provider fores the user to set a condition_scope, which
is not a valid attribute on all alert condition types.  The apm_app_metric types
require it to be set, but the API throws back a message that we surface to the
user already.  Without this change though, users of other condition types, like
apm_kt_metric, are unable to get a clean run, because the condition_scope is not
valid for that type.