fix(changelog): ensure proper branch to base from
Without this change, the pull-request action doesn't have a valid base due to
the detached head state when the tag is checked out.  Here we use the upload and
download artifact actions around switch to the master branch before we submit a
pull request.  This should result in the desired behavior, accounting for the
needs of the pull-request action.