fix(alerts): avoid bad index reference
Without this change, cases where the configured number of terms differ from that
in the API, we incorrectly reference an index that doesn't exist.  This also
looks to be the condition when we deprecate an item, and the state has
half-moved over to the new values.  Here we ensure that we check the length of
the conditions before we attempt to use the values and avoid a bad index
reference.

fixes: #882