fix(nrql_alert_condition): validate operator based on condition type
fixes #1157

Without this change, we attempt to set an invalid operator on baseline and
outlier conditions.  This looks to be a change in the upstream API to perform
more validation, which is then surfaced as an API error to the Terraform user,
but without the needed context about why the call was invalid.  Here we catch
this condition during the expandNrqlConditionTerm code to ensure that we don't
attempt to make the call, and give back a reasonable error to the user.