fix: stop fuzzy search on resource_newrelic_application_settings failing apply
https://github.com/newrelic/terraform-provider-newrelic/issues/2013

Use the same logic as data_source_newrelic_application for finding the correct APM app
in resource_newrelic_application_settings.

This resolves the issue of the resource failing to find apps that  have a name
that is a substring of another app. E.g. "api" would throw an error for too many
entities found if another app was called "privateapi".