fix(workflows): stop silently removing channels on workflow updates/deletes
Currently, if we delete a workflow or update the list of channels it uses, the provider will automatically delete the channels that are no longer used.
This behaviour goes against users' IaC intuition because the deleted channels are **still defined** in the TF config and lead to non-empty TF plan.

This PR makes the behaviour more intuitive: if you delete workflow without removing its channel from your config, the channel remains until you remove it from your config as well.
Same goes for workflow updates that remove some of the channels from workflow configuration.