feat(data/account): Allow reading the provider's `account_id`.
Right now there doesn't look to be a way to determine what `account_id` is
being used by default with the provider. Right now
`newrelic_api_access_key` requires specifying the `account_id` but there
isn't an easy way to get this information without it being passed along
correctly.

With this change, the `newrelic_account` resource would default to the
`account_id` of the provider allowing easy access to this information in
contexts that might not have easy access (nested modules).